pyproject.toml
[tool.poetry]
name = "cybulde-data-preparation"
version = "0.1.0"
description = "Data preparation for cyber-bullying model"
authors = ["abhipreets2 <abhipreets2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
hydra-core = "~=1.3.2"
pydantic = "~=1.10.7"
pandas = "~=2.0"
nltk = "~=3.8"
symspellpy = "~=6.7"
fsspec = {version = "~=2023.5", extras = ["gcs"]}
gcsfs = "~=2023.5"
google-cloud-secret-manager = "~=2.16"
fastparquet = "~=2023.4"
dvc = {version = "~=2.56", extras = ["gdrive", "gs"]}
dask = {version = "~=2023.5", extras = ["complete"]}
distributed = "~=2023.5"
dask-cloudprovider = {version = "2022.10", extras=["gcp"]}
dask-ml = "~=2023.3"
bokeh = "~=3.1"
fastapi = "~=0.104"
uvicorn = {version = "~=0.24", extras=["standard"]}

# Tokenization
tokenizers = "~=0.13"
transformers = "~=4.30.1"

[tool.poetry.group.dev.dependencies]
pytest = "~=7.3"
black = "~=23.3"
isort = "~=5.12"
flake8 = "~=6.0"
mypy = "~1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
